Update put order, cart, add cart

+++ SWP391_PreCookingPackage/Controllers/OrdersController.cs
@@ -26,18 +26,19 @@ namespace SWP391_PreCookingPackage.Controllers
 
         // GET: api/Orders
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<Order>>> GetOrders()
+        public async Task<ActionResult<IEnumerable<OrderModel>>> GetOrders()
         {
           if (_context.Orders == null)
           {
               return NotFound();
           }
-            return await _context.Orders.ToListAsync();
+            var orders = await _context.Orders.ToListAsync();
+            return Ok(_mapper.Map<List<OrderModel>>(orders));
         }
 
         // GET: api/Orders/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<Order>> GetOrder(int id)
+        public async Task<ActionResult<OrderModel>> GetOrder(int id)
         {
           if (_context.Orders == null)
           {
@@ -49,9 +50,12 @@ namespace SWP391_PreCookingPackage.Controllers
             {
                 return NotFound();
             }
-
-            return order;
+            OrderModel result = _mapper.Map<OrderModel>(order);
+            var items = _mapper.Map<List<OrderItemModel>>(_context.OrderItems.Where(oi => oi.OrderId == result.Id).Include(oi => oi.Package).ToList());
+            result.Items = items;
+            return Ok(result);
         }
+
         [HttpGet("GetByUser/{id}")]
         public async Task<ActionResult<IEnumerable<Order>>> GetOrderByUser(int id)
         {
@@ -68,17 +72,49 @@ namespace SWP391_PreCookingPackage.Controllers
 
             return Ok(orders);
         }
+        [HttpGet("GetCartByUser/{id}")]
+        public async Task<ActionResult<Order>> GetCartByUser(int id)
+        {
+            if (_context.Orders == null)
+            {
+                return NotFound();
+            }
+            var order =  _context.Orders.FirstOrDefault(x => x.UserId == id && x.Status == "On-cart");
+
+            if (order == null)
+            {
+                OrderCreateModel default_cart_order = new OrderCreateModel
+                {
+                    UserId = id,
+                    OrderDate = DateTime.Now,
+                    TotalPrice = 0,
+                    Status = "On-cart",
+                    PaymentMethod = "COD",
+                };
+                Order default_order = _mapper.Map<Order>(default_cart_order);
+                _context.Orders.Add(default_order);
+                await _context.SaveChangesAsync();
+                order = _context.Orders.FirstOrDefault(x => x.UserId == id && x.Status == "On-cart");
+            }
+            OrderModel result = _mapper.Map<OrderModel>(order);
+            result.Items = _mapper.Map<List<OrderItemModel>>(_context.OrderItems.Where(oi => oi.OrderId == result.Id).ToList());
+            return Ok(order);
+        }
 
         // PUT: api/Orders/5
         // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutOrder(int id, Order order)
+        public async Task<IActionResult> PutOrder(int id, OrderCreateModel model)
         {
-            if (id != order.Id)
+            if (id != model.Id)
             {
-                return BadRequest();
+                return BadRequest("Id doesn't match!");
             }
-
+            if(_context.Orders == null)
+            {
+                return Problem("The Orders is null!");
+            }
+            Order order = _mapper.Map<Order>(model);
             _context.Entry(order).State = EntityState.Modified;
 
             try
@@ -89,7 +125,40 @@ namespace SWP391_PreCookingPackage.Controllers
             {
                 if (!OrderExists(id))
                 {
-                    return NotFound();
+                    return BadRequest("Order not found!");
+                }
+                else
+                {
+                    throw;
+                }
+            }
+
+            return NoContent();
+        }
+
+        [HttpPut("addCart/{id}")]
+        public async Task<IActionResult> PutOrder(int id, OrderItemModel model)
+        {
+            if (id != model.OrderId)
+            {
+                return BadRequest("Order Id doesn't match!");
+            }
+            if(!_context.Orders.Any(o => o.Id == id))
+            {
+                return BadRequest("Order not found!");
+            }
+            model.Id = null;
+            OrderItem orderItem = _mapper.Map<OrderItem>(model);
+            _context.OrderItems.Add(orderItem);
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if (!OrderExists(id))
+                {
+                    return BadRequest("Order not found!");
                 }
                 else
                 {
@@ -105,15 +174,20 @@ namespace SWP391_PreCookingPackage.Controllers
         [HttpPost]
         public async Task<ActionResult<Order>> PostOrder(OrderCreateModel model)
         {
-          if (_context.Orders == null)
-          {
-              return Problem("Orders Dbset is null.");
-          }
+            if (_context.Orders == null)
+            {
+                return Problem("Orders Dbset is null.");
+            }
+            if(!_context.Users.Any(u => u.Id == model.UserId))
+            {
+                return BadRequest("User not found!");
+            }
             Order order = _mapper.Map<Order>(model);
             _context.Orders.Add(order);
             await _context.SaveChangesAsync();
-            return Ok();
-            //return CreatedAtAction("GetOrder", new { id = order.Id }, order);
+            Order result = _context.Orders.OrderByDescending(o => o.Id).FirstOrDefault();
+            //return Ok();
+            return CreatedAtAction("GetOrder", new { id = result.Id }, result);
         }
 
         // DELETE: api/Orders/5
@@ -129,7 +203,10 @@ namespace SWP391_PreCookingPackage.Controllers
             {
                 return NotFound();
             }
-
+            if(_context.OrderItems.Any(oi => oi.OrderId == id))
+            {
+                return BadRequest("Fail to delete order. The order contains item(s).");
+            }
             _context.Orders.Remove(order);
             await _context.SaveChangesAsync();
